I started addressing the issue I created 2 years ago about this shit being unreadable. Hopefully this makes the thing more clear and readable.
Some improvements could be done to the upload_file_* thing. As there's
really no need to pass argc and argv to those. And passing argv to
that is an overkill

Solution: To create a const char* variable and save the filename in
there and pass that to the functions. Simple as that. But i'm a lazy
motherfucker and I'm tired. Boss.